# demoCloth/src/main/resources/apiDefinition/DemoClothApiDefinition.yml
openapi: 3.0.3
info:
  title: Swagger DemoCloth - OpenAPI 3.0
  description: |-
    This is a sample Cloth Store Server based on the OpenAPI 3.0 specification.  
    
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.2
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: https://localhost:8080/api/v3
tags:
  - name: cloth
    description: Information details about any cloth at the store.
    externalDocs:
      description: Find out more
      url: http://swagger.io
paths:
  /cloth:
    get:
      tags:
        - cloth
      summary: Finds Cloth by date, product_id, brand_id.
      description: Find if any item of the product is available at that date, at that store 
      operationId: findItem
      parameters:
        - name: date
          in: query
          description: Date provided to check item availability
          required: true
          explode: true
          schema:
            type: string
            default: 2020-06-14-23.59.59
        - name: product_id
          in: query
          description: Identifier of the product
          required: true
          explode: true
          schema:
            type: string
            default: 35455
        - name: brand_id
          in: query
          description: Identifier of the brand
          required: true
          explode: true
          schema:
            type: string
            default: 1
  
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '204':
          description: No product available for that date, store and product
        '400':
          description: Invalid data provided
      security:
        - petstore_auth:
            - write:item
            - read:item
components:
  schemas:
    Item:
      type: object
      properties:
        product_id:
          type: integer
          format: int64
          example: 35455
        brand_id:
          type: integer
          format: int64
          example: 10
        price_list:
          type: integer
          format: int32
          example: 7
        start_date:
          type: string
          format: date-time
          example: 2020-06-14-00.00.00
        end_date:
          type: string
          format: date-time
          example: 2020-06-14-00.00.00
        price:
          type: string
          example: 25,45€
      xml:
        name: item
  requestBodies:
    Item:
      description: List of items
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Item'
        application/xml:
          schema:
            $ref: '#/components/schemas/Item'
 